auth-service/requirements.txt
Flask==3.0.0
python-keycloak==2.10.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0

